tg_bot/requirements.txt
requests==2.28.1
python-dotenv==0.21.0
ollama==0.2.0
loguru==0.6.0
python-telegram-bot==20.0
